feat: Desarrollo del tercer endpoint

+++ src/main/java/dev/franciscolorite/pruebatecnicabcnc/repository/AlbumRepository.java
@@ -1,8 +1,7 @@
 package dev.franciscolorite.pruebatecnicabcnc.repository;
 
 import dev.franciscolorite.pruebatecnicabcnc.model.Album;
-import dev.franciscolorite.pruebatecnicabcnc.model.Photo;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface AlbumRepository extends ListCrudRepository<Album, Integer> {
+public interface AlbumRepository extends ListCrudRepository<Album, Long> {
 }
